requirements.txt
antlr4-tools
black
cfpq-data
pre-commit
pydot
pyformlang~=0.1.26
pytest
scipy
